+++ src/components/CampaignDetailTable.js
@@ -58,7 +58,8 @@ export default class CampaignDetailTable extends Component {
         
         return (
             <div className="container">
-                <div className="col-md-12 col-sm-12 col-middle table-responsive animated card-item">
+                <div className="campaign-stats-table col-md-12 col-sm-12 col-middle 
+                    table-responsive animated card-item">
                     <table className="table table-hover stats-list">
                         <thead>
                             <tr>
@@ -67,7 +68,7 @@ export default class CampaignDetailTable extends Component {
                                     <h4> Views </h4> </th>
                                 <th className={`text-center ${classClicks}`}> 
                                     <h4> Clicks </h4> </th>
-                                <th className={`text-center ${classInstalls}`}> 
+                                <th className={`hidden text-center ${classInstalls}`}> 
                                     <h4> Installs </h4> </th>
                             </tr>
                         </thead>
@@ -88,7 +89,7 @@ export default class CampaignDetailTable extends Component {
                                             {numsFormatter(row.clicks)} </h4>
                                             <span className="small">Clicks</span>
                                         </td> 
-                                        <td className={classInstalls}> 
+                                        <td className={`hidden ${classInstalls}`}> 
                                             <h4 className="stats-value stats-col-installs">
                                             {numsFormatter(row.installs)} </h4>
                                             <span className="small">Installs</span>